Refactor: Create IQuestionnaire for interfaces
